refactor(server): extract mysql protocol helpers into utility module
