+++ packages/server/src/services/mysqlProtocolServer.ts
@@ -17,9 +17,15 @@ import {
   formatValueByType,
   type MySQLColumnDefinition,
 } from './mysqlResultFormatter';
+import {
+  compactSqlForLog,
+  describeCapabilityFlags,
+  doubleSha1,
+  formatCapabilityFlags,
+  verifyToken,
+} from './mysqlProtocolUtils';
 import config from '../config';
 import logger from '../logger';
-import * as crypto from 'crypto';
 import * as path from 'path';
 
 // mysql2 is CommonJS — use require for the server-side API
@@ -135,39 +141,6 @@ try {
 }
 
 /* ------------------------------------------------------------------ */
-/*  MySQL auth_41 helpers (inlined to avoid internal mysql2 imports)    */
-/* ------------------------------------------------------------------ */
-
-function sha1(...buffers: Buffer[]): Buffer {
-  const hash = crypto.createHash('sha1');
-  for (const buf of buffers) hash.update(buf);
-  return hash.digest();
-}
-
-function xor(a: Buffer, b: Buffer): Buffer {
-  const result = Buffer.allocUnsafe(a.length);
-  for (let i = 0; i < a.length; i++) result[i] = a[i] ^ b[i];
-  return result;
-}
-
-function doubleSha1(password: string): Buffer {
-  return sha1(sha1(Buffer.from(password)));
-}
-
-function verifyToken(
-  publicSeed1: Buffer,
-  publicSeed2: Buffer,
-  token: Buffer,
-  doubleSha: Buffer,
-): boolean {
-  const seed1 = publicSeed1.slice(0, 8);
-  const seed2 = publicSeed2.slice(0, 12);
-  const hashStage1 = xor(token, sha1(seed1, seed2, doubleSha));
-  const candidateHash2 = sha1(hashStage1);
-  return candidateHash2.compare(doubleSha) === 0;
-}
-
-/* ------------------------------------------------------------------ */
 /*  MySQL client capability flags (from protocol spec)                 */
 /* ------------------------------------------------------------------ */
 
@@ -184,33 +157,6 @@ const CLIENT_PLUGIN_AUTH = 0x00080000;
 const CLIENT_CONNECT_ATTRS = 0x00100000;
 const CLIENT_PLUGIN_AUTH_LENENC_CLIENT_DATA = 0x00200000;
 
-function formatCapabilityFlags(flags: number): string {
-  return `0x${flags.toString(16).padStart(8, '0')}`;
-}
-
-function describeCapabilityFlags(flags: number): string[] {
-  if (!flags) return [];
-
-  const knownFlags: Array<[string, number]> = [
-    ['LONG_PASSWORD', CLIENT_LONG_PASSWORD],
-    ['FOUND_ROWS', CLIENT_FOUND_ROWS],
-    ['LONG_FLAG', CLIENT_LONG_FLAG],
-    ['CONNECT_WITH_DB', CLIENT_CONNECT_WITH_DB],
-    ['NO_SCHEMA', CLIENT_NO_SCHEMA],
-    ['PROTOCOL_41', CLIENT_PROTOCOL_41],
-    ['TRANSACTIONS', CLIENT_TRANSACTIONS],
-    ['SECURE_CONNECTION', CLIENT_SECURE_CONNECTION],
-    ['MULTI_RESULTS', CLIENT_MULTI_RESULTS],
-    ['PLUGIN_AUTH', CLIENT_PLUGIN_AUTH],
-    ['CONNECT_ATTRS', CLIENT_CONNECT_ATTRS],
-    ['PLUGIN_AUTH_LENENC_CLIENT_DATA', CLIENT_PLUGIN_AUTH_LENENC_CLIENT_DATA],
-  ];
-
-  return knownFlags
-    .filter(([, value]) => (flags & value) !== 0)
-    .map(([name]) => name);
-}
-
 /* ------------------------------------------------------------------ */
 /*  Connection state                                                   */
 /* ------------------------------------------------------------------ */
@@ -223,12 +169,6 @@ interface ConnectionState {
   connectedAt: Date;
   lastActivity: Date;
   idleTimer: ReturnType<typeof setTimeout> | null;
-}
-
-function compactSqlForLog(sql: string, maxLen = 300): string {
-  const compact = sql.replace(/\s+/g, ' ').trim();
-  if (compact.length <= maxLen) return compact;
-  return `${compact.substring(0, maxLen)}...`;
 }
 
 /* ------------------------------------------------------------------ */
@@ -670,14 +610,3 @@ export class MySQLProtocolServer {
 }
 
 export default MySQLProtocolServer;
-
-// Test-only exports for unit testing pure helper functions
-export const __testHelpers = {
-  sha1,
-  xor,
-  doubleSha1,
-  verifyToken,
-  formatCapabilityFlags,
-  describeCapabilityFlags,
-  compactSqlForLog,
-};
